Adding ONIG_HOME environment variable check
Setting the ONIG_HOME env variable to the path of a local, pre-built, copy of libonig will now allow the download & build steps for onig to be skipped. This is useful for building on Windows where some of the build steps will not work directly (ex: gzip which is usually not installed on Windows, and system calls expecting a bash shell for execution).